BottomAppbar, FAB を MainActivity に移動
