+++ app/src/main/res/layout/activity_main.xml
@@ -32,7 +32,27 @@
             android:id="@+id/viewPager"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
+            android:paddingBottom="?actionBarSize"
             app:layout_behavior="@string/appbar_scrolling_view_behavior" />
 
+        <com.google.android.material.floatingactionbutton.FloatingActionButton
+            android:id="@+id/fab"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginEnd="16dp"
+            android:layout_marginBottom="16dp"
+            android:contentDescription="@string/input_content"
+            android:src="@drawable/ic_add_24px"
+            app:layout_anchor="@id/bottomAppBar" />
+
+        <com.google.android.material.bottomappbar.BottomAppBar
+            android:id="@+id/bottomAppBar"
+            style="@style/Widget.MaterialComponents.BottomAppBar"
+            android:layout_width="match_parent"
+            android:layout_height="?actionBarSize"
+            android:layout_gravity="bottom"
+            app:fabCradleMargin="0dp"
+            app:menu="@menu/menu_main_bottom_app_bar"
+            app:navigationIcon="@drawable/ic_more_vert_24px" />
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
 </layout>
